Fix architecture
